requirements.txt
numpy==1.21.6
numba==0.55.1
matplotlib==3.3.4
scipy==1.8.0
statsmodels==0.13.5
pytorch-lightning==1.9.5
pytorch-forecasting==0.10.3
sympy==1.12.1
reformer_pytorch==1.4.4
openpyxl==3.1.5
einops==0.8.0